requirements/requirements-dev.txt
# 🛠️ 개발 환경용 추가 패키지들

# 코드 품질 도구
black>=23.0.0                 # 코드 포매터
flake8>=6.0.0                 # 린터
isort>=5.12.0                 # import 정렬
mypy>=1.0.0                   # 타입 체커

# 테스트 도구
pytest>=7.0.0                 # 테스트 프레임워크
pytest-cov>=4.0.0             # 테스트 커버리지
pytest-mock>=3.10.0           # 모킹 도구

# 문서화 도구
sphinx>=6.0.0                 # 문서 생성
sphinx-rtd-theme>=1.2.0       # 문서 테마

# 개발 유틸리티
ipython>=8.0.0                # 개선된 Python REPL
jupyter>=1.0.0                # Jupyter 노트북
notebook>=6.5.0               # Notebook 서버

# 성능 분석
memory-profiler>=0.60.0       # 메모리 프로파일링
line-profiler>=4.0.0          # 라인별 성능 프로파일링

# 디버깅 도구
pdb++>=0.10.3                 # 개선된 디버거
icecream>=2.1.0               # 디버깅 출력

# 추가 Scrapy 관련 도구
scrapy-splash>=0.8.0          # JavaScript 렌더링
scrapy-fake-useragent>=1.4.0  # User-Agent 로테이션
scrapy-rotating-proxies>=0.6.2 # 프록시 로테이션

# 데이터 처리
pandas>=2.0.0                 # 데이터 분석
matplotlib>=3.7.0             # 그래프 생성
seaborn>=0.12.0               # 통계 시각화

# API 개발 (선택사항)
fastapi>=0.100.0              # 웹 API 프레임워크
uvicorn>=0.22.0               # ASGI 서버
